ocr_service/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.27.0
google-genai>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9
pydantic-settings>=2.2.0
Pillow>=10.0.0
